requirements.txt
numpy>=1.20.0
scipy>=1.7.0
pandas>=1.3.0
pyyaml>=6.0
tqdm>=4.60.0
pyarrow>=10.0.0
